+++ A2lParser/Interfaces/IXcpEvent.cs
@@ -11,7 +11,7 @@ namespace A2lParserLib.Interfaces
         public string ShortName { get; set; }
         public byte Channel { get; set; }
         public int MaxDaqList { get; set; } //Maximum number of DAQ lists in this event channel
-        public byte TimeCycle { get; set; } //Event channel time cycle
+        public uint TimeCycle { get; set; } //Event channel time cycle
         public byte TimeUnit { get; set; } //Event channel time unit
         public uint Ident { get; set; } //Event channel identifier, some A2L files have this property EVENT_CAN_ID_LIST
     }
